mongo: Change mgo package to mongo-go-driver
Since we discovered that mgo wasn't going to be maintained anymore,
change it was the main thing to do. After researching, the mongo
official driver (mongo-go-driver) seemed to be the right choice
to make this happen. This commit represents exactly this change
being done.

- Changes some query entity types, to be compatible with the new driver
- Changes the controller function that parses string to int for int64
- Changes the implementation of the data store to be compatible with
  the new driver

+++ pkg/entities/query.go
@@ -9,8 +9,8 @@ type Query struct {
 	ThingID    string
 	SensorID   string
 	Order      int
-	Skip       int
-	Take       int
+	Skip       int64
+	Take       int64
 	StartDate  time.Time
 	FinishDate time.Time
 }
